requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-dotenv
requests
xhtml2pdf